CDD-2418: Upgraded to play2.6 (#29)

+++ app/uk/gov/hmrc/customs/notification/receiver/repo/NotificationRepo.scala
@@ -30,9 +30,9 @@ trait NotificationRepo {
 
   def notificationsByConversationId(conversationId: ConversationId): Future[Seq[NotificationRequest]]
 
-  def notificationCountByCsId(csid: CsId): Future[Long]
+  def notificationCountByCsId(csid: CsId): Future[Int]
 
-  def notificationCountByConversationId(conversationId: ConversationId): Future[Long]
+  def notificationCountByConversationId(conversationId: ConversationId): Future[Int]
 
   def notificationCount: Future[Int]
